Try to make sure the group dialog stays within a sane width limit.
The approach taken is "behave reasonably on 640x480 and above" by
clipping the bounding box of the task items to the max() of that
minimum supported resolution (taking dialog margins and, if vert-
ically oriented, a likely panel into account) and the applet's
width.

A perfect impementation would figure out the dimensions of the
work area (i.e. screen sans relevant struts) and clip to that,
but knowing how much C++ that takes to do correctly on multi-
head (as Yakuake does it) and seeing how the goal is to have a QML
task manager, I don't feel that's maintainable for now. Let's look
into improving PlasmaCore.Dialog in libplasma2 instead to help
with avoiding too unreasonably large dialogs instead, or perhaps
Qt Quick 2 will evolve a helper. There's a TODO in there for now.

On the flip side, this actually behaves better than the old
applet anyway, which had rather more unreasonable popup sizing ...

BUG:322248