Sizing threshold tweaks.
This reduces the preferred minimum height of items to barely above
the font's M size, allowing layout to break into additional rows
far sooner. Additionally, the scaling curve for icons is fudged to
allow the icon to overrun a bit into the frame margins at very low
item heights. Together this answers the clamor for supporting very
space-constrained forced multi-row configurations similar to the
old applet's behavior, while still trying to look not quite as bad
in the process.

Not very happy about introducing some fixed pixel values into the
calculations, but between the poor support for font metrics in QML
and tweaking things for scenarios in which single pixels matter,
it's inevitable for now.

BUG:321849