Improve finding of X11 in main CMakeLists.txt
Q_WS_X11 should not be defined on a Qt 5 only build which meant that we
never looked for X11 in the first place as the find package for X11 was
only in the if (Q_WS_X11).

X11 is now found unconditionally and considered as an optional build
dependency. If it is found the normal checks are continued which add
required dependencies for the X11 build of kde-workspace.

Thanks to Kubuntu and project neon 5 for finding this build issue.