crashes--: Rework applet handle anchoring
Use a property to track the parent item. This makes it easier to do the
re-anchoring of the applet handle on size changes. One effect is that
property bindings are apparently triggered in a more deterministic way.
A nice side effect is that this simplifies the code a bit.
Lesson: be *very* careful when swapping out individual properties in
javascript code, rather rely more on property bindings to do the work
for you.

IOW: This fixes a nasty crasher in the applet handle.

CCMAIL:plasma-devel@kde.org