Cursor Theme KCM: Show correct resize cursor in preview for themes without a file called "size_fdiag"
Apparently in XCursorTheme::findAlternative() (file
kcontrol/input/xcursor/xcursortheme.cpp) the alternatives for
"size_bdiag" and "size_fdiag" are swapped, so for themes not containing
"size_fdiag" the wrong resize cursor is shown in the preview.

This patch fixes that long standing bug. (there has been no change to
that function since 2007!)

This also fixes the glitch mentioned in bug#325763, that the wrong
arrows are used for the window resize hint after the theme change is
applied (for the current X session).

CCBUG: 325837