[khotkeys/kcm] Drop workaround for a Qt bug
It's not possible to figure out why this workaround was needed as
the given bug url is invalid and the bug ID is not valid in current
Qt bug tracker.

Let's assume that this problem is fixed and we can drop the code.
If it's not fixed we have the git history to restore.

Primary reason to drop the workaround is the usage of Q_PRIVATE_SLOT
which doesn't want to compile any more with Qt5.