[kcm-cursortheme] Port away from QCursor - XCursor magic
We cannot create a QCursor from an X cursor handle any more. So we
need to go a little bit more low-level for setting the cursor on the
preview widget.

Loading a theme does no longer return a QCursor but the handle and
the PreviewCursor has to destroy the handle in it's dtor. Of course
only done if the platform is X11.