We mourn today the death of krandr
The young lord kscreen supported by General Fiestas and Vrátil started
a revolution. First they took over the independent republic Kubuntu in
the land of Ubuntu and the city where all the gear heads live in the
land where the people wear red fedoras. Over time the young lord
got control over all the gear heads and his people loved him. But still
the old king krandr kept his place in his palace.

Now came the day when the supporters of the young lord finally entered
the palace in the capital and stabbed the old king. With his last words
the king asked "Et tu, Alex?" and Alex cried "but, I love you!".

Immediately kscreen got proclaimed to the new king of screen management.
And the people in the streets celebrated their new ruler:

THE KING IS DEAD, LONG LIVE THE KING!