Delete workspaceoptions KCM for the time being
This KCM is not useful in this specific way for Plasma2. The options
will be brought back, but at better matching places. Like switching
form factors needs to go with other form factor related options. The
dashboard settings should go into the dashboard and the tooltip
setting belongs to look and feel.