A dream comes true: the screensaver kcm goes away
As we are not going to have XScreenSaver based screen savers in Plasma 2
anymore this KCM is not very useful any more. There are some config
options which need to be preserved, but it doesn't make any sense to try
to split them out of this extremly old and complex code base. It's better
to just reimplement them.