Port KSplashQML to Qt5 and remove its XLib code
Martin G. suggested to try if it would simply work without any XLib/XCB
bits at all and it does work in testing mode, I'm unsure if it also
works after actual login, but I don't see why it wouldn't.

KSplash was also receiving the stage messages via XAtoms, which I
replaced with simple DBus interface. That means that all the parts in
the login sequence need to be updated (I'll do it) to call a dbus method
instead of sending X message. Same API is used as the XAtoms were using,
ie. sending strings of component names. In some future I'd probably also
change the stages resolution as it's not too robust and does not help
paralelization much. I have a plan in my head but that's for another
commit.

REVIEW: 113415