Bookkeeping for tasks
This makes migration of task objects between shown and hidden work.
Depending on their status, and the shown property of the task, the task
is moved between the lists when its status updates. This commit also
fixes forwarding the status in the PlasmoidInterface, which wasn't wired
up yet.

Memory management has changed a bit, too, we now keep references to the
declarative lists in the private object.

In order to make shown / hidden preferences for tasks work, we need to
book-keep this information in host, not in the individual tasks (as
these are global objects, and we want to change configuaration per host.