Systemtray: Delegate Sizing
This commit makes the gridview more dynamically sized. I think in a
world of Retina displays, we should move away from pixel-based sizes as
much as possible. From this point on, protocols are simply not allowed
to determine the size of their items. We can still apply tricks in the
possible x11 protocol, and have support for images and movies of fixed
size working well. The input area absolutely has to be client-defined,
we can't have applications tell us "the interactive area for this app is
22x22 pixels" -- that is not going to happen, as it simply breaks on
different dpi *and* input devices.

As such, the item size is now more dynamic, the logic is not complete
yet, but the direction of this gets clearer. This makes the "basic"
itemSize a "recommendation" to the UI for now (it will become an even
more internal detail later).

Otherwise, the system tray is coming along nicely. What works:

- Display of status notifier items (not interactive yet, just icons)
- Embedding of plasmoids such as battery, network
- Status: items are shown and hidden (moved between the models, really)

A few things that I want to improve at a later stage is the messy popup
situation. We can probably embed the popups of the plasmoids and
statusitems into the systemtray popup and streamline this navigational
process a bit, while keeping the same interaction model.

We're obviously not feature-complete yet and it's all very
work-in-progressish, but it's "coming along nicely". :)

Here's a screenshot: http://i.imgur.com/LuQdkDL.png

CCMAIL:plasma-devel@kde.org
DIGEST: