More repaint and powerfriendly needsattention animation
This animation is visually almost the same as the one it replaces, but
saves 80+% of the repaints, which can lead to large areas being
repainted.

The trick here is instead of using a curve that is flat at the start
(but still leads to repaints) by sleeping at first, and then
accelerating much faster, thereby emulating the more rapid acceleration
of the longer animation at the end of the first animation.

Positive side-effect: By moving the pause animation at the end of the
first cycle we now react almost half a second faster to the change in
status.