[oxygen] Set path to color scheme as XProperty to toplevel windows
An application can indicate the color scheme it is using through the
dynamic property "KDE_COLOR_SCHEME_PATH" on the QApplication. The style
can forward this path on all toplevel windows to inform the window
manager about it. On X11 this is done through the
_KDE_NET_WM_COLOR_SCHEME X property.

The property gets set when the window is first shown and updated or
removed when the palette changes for the toplevel window.

Once there is a new KStyle again this functionality should move into
the more generic style to not have to copy the functionality into each
style.