Delay calendar initialization in digital clock
We use a Loader now to encapsulate the whole Calendar shebang, saving
hugely in startup time. Also, as the calendar is really only loaded
on-demand, it won't use any memory if you don't use it. This also means
that we can prevent Akonadi from autostarting on login, at least as far
as the Calendar is concered.

Unrelated: CompactRepresentation is now called DigitalClock.qml, as
that's exactly what it is.

DIGEST: lazy loading of Plasma Calendar