Build without QImageBlitz available
Just got fed up of this module never building for me. We don't really
use QImageBlitz in there anyway, if it was the case someone would have
noticed I guess as it got no Qt5 port.

So pretend it's optional (I suspect we'll just ditch it completely later
on, it's used only twice and in parts we don't build for now).