KInfocenter/OpenGL: fix ReadPipe() to return 0 in the case that the command cannot be run
ReadPipe() doesn't return 0 as expected in the case that the command is
not found. but the length of sh's output which is "command not found"
in this case. This is because popen() does not fail if the command is
not found, because it _can_ run "sh". (according to the man page, popen
calls "/bin/sh -c command")
To fix this, ReadPipe() should check the return code of the call to
pclose() (see "man pclose"), and return 0 if this is not equal to 0.

CCBUG: 327382
REVIEW: 113779