[klipper] Remove tearOffHandling in klipperpopup
The functionality got removed in KGlobalSettings with git commit
a18e1843ca2666894b4db75ca3be07ae3b3d3770, because there is no GUI
to set it and no application except klipperpoup even looked at it.

The result is that due to lacking GUI this had always been disabled
in kde4 times, so we can just remove it.