Improve locale time formatting in Digital Clock applet
Qt's QLocale does not offer any modular time creating like Klocale did,
eg. no "gimme time with seconds" or "gimme time without seconds and with
timezone".

QLocale supports only two formats - Long and Short. Long is unusable in
many situations and Short does not provide seconds. So if seconds are
enabled, we need to add it manually to the Short format that QLocale
provides us.

That's done by JS regexp replace that looks for the delimiter between
"h" and "m", takes it and appends it after "mm" and then appends "ss"
for the seconds. Also it checks if the format string already does not
contain the seconds part.

Reviewed-by: Sebastian Kügler