Reinstate timezone support in Digital clock applet
Appends the timezone string to the end. This might be wrong for some
Locales, however QLocale does not provide any hint about where the
timezone string should be, so we're placing it always at the end.

Looking at all Qt's Locale time formats, only zh_* locales do it
differently - some of them prepend it, some of them put it in square
brackets, eg. [t]. An exception for zh_* could be added maybe.