[screenlocker] Make LockWindow a more normal QWidget
The QWidget::create method does no longer create the QWidget from the
native window, so just create the QWidget as override redirect. Our
custom created window did nothing in addition.

Something with the rendering is wrong. Qt never calls the paint event.
Because of that the background pixel is set to black.