Recalculate totalAmount when possible
When the job is picked up while running, totalAmount will be unset, as
this is only emitted by signal on startup of the job, but not updated
all the time (sensibly). If we miss the job appearing for some reason,
we can still do an estimated guess of the total amount.