Fix applet ordering in panel
The call to LayoutManager.save() would silently fail, while it works
just fine referencing it locally. I'm not sure if that's a peculiarity
in the new v4 engine, or just bad usage of namespaces on our side, but
as this solution seems both, correct code and working code, let's go for
it.