[ksplash] Make KPlashQML stages ready to be run asynchronously
Since we ported to DBus we are using ASync calls to send the
QDBusMessages that indicate to KSplash that a new stage has happened.
Since we are using an async method we can't take into account the order
anymore.

I tried changing all calls to be sync and then KSplash works correctly
which shows that we are starting the session correctly but it is better
if we keep these calls async because you know, the future is async.

CCMAIL: plasma-devel@kde.org