Revert cae28373 as it crashes
This commit introduced a crasher on my system, somewhere deep in v4 with
an utterly useless backtrace. I'm not sure which bug it fixes, but a
shell crash is certainly worse (and hard to say why it crashes, it
shouldn't).

Eike, sorry for reverting your commit, we'll have to investigate why it
crashes, and if we can solve the problem in a way that doesn't.

CCMAIL:hein@kde.org