Fixed width for digital clock
With this commit, the clock is less jumpy when the time changes. We now
depend on a fixed string, that has the length of the time format, but
uses A for an just above average sized width. This string is used both
to compute the available font sizes, and for the width computation of
the widget in the panel.

This panel relayouts when the time changes.