[oxygen] drawFloatFrame for QtQuick menus
drawFrameMenuPrimitive is entered with a null QWidget if a QQuickMenu is
rendered. Thus it doesn't do anything.  By checking whether it is a
QQuickItem we get the menu to render with a frame. Unfortunately we
cannot render the widget background as that really needs a non-null
QWidget and we also do not yet get shadows.

REVIEW: 114271