Add default OSD to our default look&feel package
The OSD features just an icon and (a progress bar OR a text). It uses
standard Plasma components for drawing things, Marco has on his todo
list to turn the progress bar black, will look better when that's done.

When the OSD is showing progress, no text is shown (that's part of the
QML implementation and can be changed by the look&feel package).

REVIEW: 115260