Make WidgetExplorer load super fast
Use of PlasmaComponents.PageStack had some problems.
Changing the mainItem in a dialog causes an expensive sync in Dialog
We make the dialog visible before the everything is loaded

It had a lot of overhead and slow syncronous methods that didn't really
add anything to the class.

This makes everything instant with no obvious regressions.