width and height properties for Wallpaper.Image
These properties are needed to set the targetSize in the Image that is
instantiated from the configuration. The values for the preferred
targetsize need to be set by the wallpaper QQuickItem.

This is the C++ portion of the fix for picking the right wallpaper, and
thus preventing ugly scaling effects.