event compress layout size update and orientation
event compressing both the resize of the panel and the change of the formfactor update in the layout, makes it way faster and reliable.
before due to the non deterministic order of signal/slot invocation, sometimes the shell got blocked for minutes (or even forever) at 100% cpu.
this doesn't seem to happen anymore