[kinfocenter] Add new module for a Wayland Compositor
Inspired by the module for X-Server we have a new module for Wayland
Compositors. The module introduces a new optional build dependency to
wayland client libraries.

Just like the X-Server module it talks to Wayland directly without the
help of Qt. It connects to the registry and adds all reported
interfaces to the tree widget. For outputs and seats it registers a
listener to also include this information. For every output the
information is appended, for a seat the capabilities are included.

This approach should allow to add information about further interfaces
if they provide interesting information in future.

REVIEW: 115989