Wait a second until the systray starts loading
This is a bit hacky, but it allows to load other applets in the panel
first, and gives them some time to do so. This feels like a faster
startup of Plasma, since the panel doesn't stay empty now until the
whole systemtray has loaded.

Let's give this a shot for others to test on their systems.