Fix compact sizing
The tooltip area, containing the small arrow pushed the last item in the
list into the next row, leading to a bit of whitespace, and a seemingly
missing icon.

The arrow might actually be rendered a bit smaller than the tooltip area
it's located in. Using the tooltip size to compute the preferred size of
the system tray fixes this.