Remove KNotifications from ksmserver
We are not going to have login/logout effects sounds
anymore, and even if we want to restore them at
some point we will not want to wait for them to finish
playing.

Also, the lock/unlock notifications are useless, so
removing KNotifications from ksmserver in Plasma Next
is not a drama.