Fix deadlock in KSMServer::runUserAutostart()
KSMServer before this commit was launching autostarted desktop files
using KRun::KRun.

KRun will end up using KInvocationTool which performs a DBus blocking
call to KLauncher in order to execute the app. This blocking call
will not return until KLauncher thinks that the app has been executed.

When KLauncher thinks that an app has been executed depends on the app
type. If they are of type DBusUnique or DBusMulti KLauncher won't
consider them executed until the approapiate things happen on the bus.

For example, a DBusUnique app will have to register a service on the
bus kind of org.kde.kded or org.kde.kwalletd.

The deadlock happens because these applications will NOT register the
dbus service until after the QSessionManager has been called, which
means that apps will try to talk to ksmsever and this will be blocked
waiting for the application to be launched (because KLauncher will
wait until the dbus service is registered).

With this commit we use QProcess::startDetached("kioclient") to
execute the desktop files which will not block since the KRun call
will happen in another process.

CCMAIL: dfaure@kde.org